pkg/metrics: Add ownerReference to metrics Service object (#1037)
* pkg/metrics: Add ownerRef to Service

This allows the Service cleanup to happen when owner is
deleted. Owner is detected by tracing up until last owner is
found. If no owner is found, the pod is set as owner.

* Gopkg.lock: Rerun dep ensure

* pkg/metrics: Update Service if it exists

When Service already exists but the OwnerRef might change,
we must make sure to the propagate those changes to the Service object
by updating the object.