Fix permission issue when settin the operator-condition
The conditions moved from the `Status` field to the `Spec`, but when setting the condition, the code still uses `client.Status().Update()` to set the condition, instead of just `client.Update()`. This is wrong in two ways: first, the code is trying to update the spec, so it will fail, and also, the operator does not have permissions to update the status. As result, the `Set` function always returns error.

This PR changes the client call to use the regular client instead of the status client.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>