Add conditions.Factory (#58)
The conditions.NewCondition function reads cluster state to determine
the namespace the operator is running in so that it can manipulate the
correct OperatorConditions object in a cluster.

However this function's dependency on a specific file existing in a
specific location makes it difficult to test code that uses it.

The Factory interface is introduced to enable users to provide their own
implementations for building a Condition object to manage OperatorCondition
resources. Users can now build a factory and inject it into code that builds
new conditions. This enables code under test to use the Factory abstraction
and enables operator authors to inject custom functionality needed when
writing tests.

Signed-off-by: Joe Lanford <joe.lanford@gmail.com>