Fix the assertMetrics test helper function
The assertMetrics function in
handler/instrumented_enqueue_object_test.go, wasn't really doing
anything. The switch-case compared a pointer value to another pointer,
so the no case was actually ever selected, because the addresses were
not the same.

Tested with a debugger - no value were ever selected.

This commit fixes this test by selecting the value rather than the
address.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>